src/pyproject.toml
[tool.poetry]
name = "sprig-poc-backend"
version = "0.0.1"
description = "PoC implementation of the Sprig back-end"
authors = ["Dorran Howell <dorran.howell@gmail.com>"]
packages = [{include = "sprig"}]

[tool.poetry.scripts]
sprig = 'sprig.app.cli:cli'

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.4"
pyarrow = "^15.0.2"
pandas = "^2.2.1"
click = "^8.1.7"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.29"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"
hypothesis = "^6.75.3"
pyright = "^1.1.308"
ruff = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
lint.select = ["F", "I"]
